+++ lib/types/moov.d.ts
@@ -175,6 +175,34 @@ export class Moov {
      */
     get capabilities(): Capabilities;
     /**
+     * Gets the Enriched Address API.
+     * @returns {EnrichedAddresses}
+     * @tag Moov
+     * @example
+     * const moov = new Moov(...);
+     * try {
+     *   await moov.enrichedAddresses.get(...);
+     * } catch (err) {
+     *   // ...
+     * }
+     */
+    get enrichedAddresses(): EnrichedAddresses;
+    _enrichedAddresses: EnrichedAddresses;
+    /**
+     * Gets the Enriched Profile API.
+     * @returns {EnrichedProfiles}
+     * @tag Moov
+     * @example
+     * const moov = new Moov(...);
+     * try {
+     *   await moov.enrichedProfiles.get(...);
+     * } catch (err) {
+     *   // ...
+     * }
+     */
+    get enrichedProfiles(): EnrichedProfiles;
+    _enrichedProfiles: EnrichedProfiles;
+    /**
      * Gets the Payment Methods API.
      * @returns {PaymentMethods}
      * @tag Moov
@@ -207,7 +235,7 @@ export class Moov {
      * @returns {Wallets}
      * @tag Moov
      *
-     * @example
+    * @example
      * const moov = new Moov(...);
      * try {
      *   await moov.wallets.get(...);
@@ -248,6 +276,8 @@ import { Capabilities } from "./capabilities.js";
 import { Transfers } from "./transfers.js";
 import { Avatars } from "./avatars.js";
 import { BankAccounts } from "./bankAccounts.js";
+import { EnrichedAddresses } from "./enrichedAddress.js";
+import { EnrichedProfiles } from "./enrichedProfile.js";
 import { PaymentMethods } from "./paymentMethods.js";
 import { Wallets } from "./wallets.js";
 //# sourceMappingURL=moov.d.ts.map