+++ CHANGELOG.md
@@ -5,11 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.1.11] - 2022-03-08
+
+## [0.1.11] - 2022-03-09
 
 ### Added
 
 - Avatars API with examples
+- Enriched Address API with examples
+- Enriched Profile API with examples
 - Updated typescript types
 - Updated documentation